requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.5.1
python-multipart==0.0.9
pillow==10.4.0
numpy==2.0.1
torch==2.3.1
torchvision==0.18.1
facenet-pytorch==2.5.3
